refactor: changes folders structure by FSD, and added mixin for future styling
