+++ src/app/styles/index.scss
@@ -1,2 +1,3 @@
 @import './reset.scss';
 @import './vars.scss';
+@import './mixin.scss';
