+++ src/pages/LayoutHome/Layout.tsx
@@ -1,13 +1,13 @@
 import React from 'react';
 import { LayoutHeader } from '../../widgets/LayoutHeader';
 import { LayoutFooter } from '../../widgets/LayoutFooter';
-import { Sections } from '../../pages/LayoutHome/ui/Sections';
+import { SectionHome } from './ui/Sections';
 
 export const Layout = () => {
   return (
     <>
       <LayoutHeader />
-      <Sections />
+      <SectionHome />
       <LayoutFooter />
     </>
   );
