+++ src/pages/LayoutHome/index.ts
@@ -1,1 +1,1 @@
-export { Sections } from './ui/Sections';
+export { SectionHome } from './ui/Sections';
