+++ src/pages/LayoutHome/ui/SectionCountry/ui/SectionCountry.module.scss
@@ -1,0 +1,73 @@
+
+
+.country {
+    background-image: url(/src/shared/assets/background/poizon.png);
+    background-color: #000000;
+    color: white;
+    padding-bottom: 2rem;
+  
+    .wrapper {
+      max-width: 1860px;
+      margin: 0 auto;
+      padding: 0 60px;
+  
+      .blocks {
+        display: flex;
+        flex-direction: row-reverse;
+  
+        .block {
+          max-width: 700px;
+          .title {
+            text-align: center;
+            padding-top: 50px;
+            font-size: 35px;
+          }
+  
+          .paragraph {
+            text-align: justify;
+            font-style: normal;
+            font-weight: 500;
+            font-size: 18px;
+            line-height: 21px;
+            color: #ffffff;
+            padding: 30px 15px 35px;
+          }
+        }
+  
+        @media screen and (max-width: 768px) {
+          .block{
+            .title {
+              padding-top: 0px;
+              font-size: 28px;
+            }
+            .paragraph {
+              padding: 10px 10px 10px 10px;
+              text-align: center;
+              font-size: 16px;
+            }
+          }   
+        }
+  
+        @media screen and (max-width: 360px) {
+          .block{
+            .title {
+              padding-top: 0px;
+              font-size: 20px;
+            }
+            .paragraph {
+              font-size: 9px;
+            }
+          }   
+        }
+      }
+      @media screen and (max-width: 1210px) {
+        .blocks{
+          // display: flex;
+          align-items: center;
+          justify-content: center;
+          flex-direction: row;
+        }   
+      }
+    }
+  }
+
