+++ src/pages/LayoutHome/ui/SectionLogistic/index.ts
@@ -1,0 +1,1 @@
+export { SectionLogistic } from './ui/SectionLogistic';
