+++ src/pages/LayoutHome/ui/SectionOrder/index.ts
@@ -1,0 +1,1 @@
+export { SectionOrder } from './ui/SectionOrder';
