+++ src/pages/LayoutHome/ui/SectionOrder/ui/SectionOrder.module.scss
@@ -1,0 +1,160 @@
+.order {
+    background-image: url(/src/shared/assets/background/green.png);
+    background-size: cover;
+    color: white;
+    padding-bottom: 5rem;
+  
+    .wrapper {
+      max-width: 1860px;
+      margin: 0 auto;
+      padding: 0 60px;
+  
+      .blocks {
+        display: flex;
+        justify-content: center;
+        // flex-wrap: wrap;
+  
+        .title {
+          padding-top: 250px;
+          font-size: 84px;
+        }
+  
+        .paragraph {
+          padding-top: 9px;
+          padding-bottom: 25px;
+          max-width: 500px;
+          font-size: 23px;
+        }
+  
+        .block {
+          max-width: 700px;
+          //height: 250px;
+          background: rgba(0, 0, 0, 0.7);
+          backdrop-filter: blur(1.54379px);
+          padding-left: 60px;
+          padding-right: 60px;
+          padding-top: 30px;
+          border-radius: 30px;
+          text-align: justify;
+  
+          h3 {
+            font-family: 'Arial';
+            font-style: normal;
+            font-weight: 700;
+            font-size: 19.949px;
+            line-height: 29px;
+            /* or 144% */
+  
+            display: flex;
+            align-items: center;
+            text-transform: uppercase;
+            padding-bottom: 9px;
+          }
+  
+          p {
+            // font-family: 'SF Pro Display';
+            // font-style: normal;
+            // font-weight: 500;
+            // font-size: 18.1383px;
+            // line-height: 24px;
+            // color: #ffffff;
+            padding-bottom: 10px;
+          }
+        }
+        .certificate {
+          padding-top: 70px;
+        }
+  
+        @media screen and (max-width: 1210px) {
+          .title {
+            text-align: center;
+            padding-top: 100px;
+            font-size: 50px;
+          }
+          .paragraph {
+            padding-top: 20px;
+            margin: 0 auto;
+          }
+          .certificate {
+            padding-top: 0px;
+          }
+        }
+  
+        @media screen and (max-width: 768px) {
+          .title {
+            padding-top: 50px;
+            font-size: 30px;
+          }
+          .paragraph {
+            padding-top: 20px;
+            text-align: center;
+            font-size: 16px;
+          }
+          .certificate {
+            padding-top: 0px;
+          }
+  
+          .block {
+            h3 {
+              font-size: 19px;
+            }
+  
+            p {
+              font-family: 'SF Pro Display';
+              font-style: normal;
+              font-weight: 500;
+              font-size: 16px;
+              line-height: 24px;
+              color: #ffffff;
+              padding-bottom: 10px;
+            }
+          }
+        }
+  
+        @media screen and (max-width: 360px) {
+          .title {
+            padding-top: 20px;
+            font-size: 20px;
+          }
+  
+          .paragraph {
+            padding-top: 20px;
+            text-align: center;
+            font-size: 12px;
+          }
+  
+          .block {
+            padding-top: 20px;
+            padding-left: 30px;
+            padding-right: 30px;
+            // max-width: 330px;
+            // height: 180px;
+  
+            h3 {
+              font-size: 10px;
+            }
+  
+            p {
+              font-size: 8px;
+              // line-height: 2;
+              color: #ffffff;
+              padding-bottom: 10px;
+            }
+          }
+        }
+      }
+    }
+  }
+  
+  @media screen and (max-width: 1210px) {
+    .blocks {
+      flex-wrap: wrap;
+    }
+  }
+  
+  @media screen and (max-width: 1024px) {
+    .order {
+      padding-bottom: 0px;
+    }
+  }
+
