+++ src/pages/LayoutHome/ui/SectionService/index.ts
@@ -1,0 +1,1 @@
+export { SectionService } from './ui/SectionService';
